# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Only Creators - Gaming Creator Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputAccountType = typeof RegisterInputAccountType[keyof typeof RegisterInputAccountType];


export const RegisterInputAccountType = {
  customer: 'customer',
  creator: 'creator',
} as const;

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  accountType: RegisterInputAccountType;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  accountType: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  role?: string;
  isBanned?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserUpdate {
  username?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  /** @nullable */
  description?: string | null;
  sampleCount?: number;
}

export interface CreatorProfile {
  id: number;
  userId: number;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  gmailAddress?: string | null;
  /** @nullable */
  discordUsername?: string | null;
  /** @nullable */
  servicesOffered?: string | null;
  /** @nullable */
  experienceLevel?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  socialLinks?: string | null;
  verificationStatus?: string;
  level?: string;
  totalLikes?: number;
  totalSamples?: number;
  followerCount?: number;
  isFeatured?: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  joinedAt?: string | null;
  createdAt: string;
}

export interface CreatorProfileInput {
  displayName: string;
  bio?: string;
  avatarUrl?: string;
  gmailAddress?: string;
  discordUsername?: string;
  servicesOffered?: string;
  experienceLevel?: string;
  portfolioUrl?: string;
  socialLinks?: string;
}

export interface CreatorProfileUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  gmailAddress?: string;
  discordUsername?: string;
  servicesOffered?: string;
  experienceLevel?: string;
  portfolioUrl?: string;
  socialLinks?: string;
}

export interface Sample {
  id: number;
  creatorId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  gameType?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  previewImageUrl?: string | null;
  /** @nullable */
  previewVideoUrl?: string | null;
  /** @nullable */
  tags?: string | null;
  status: string;
  likeCount?: number;
  commentCount?: number;
  isLiked?: boolean;
  isFavorited?: boolean;
  /** @nullable */
  creatorName?: string | null;
  /** @nullable */
  creatorAvatarUrl?: string | null;
  creatorVerified?: boolean;
  createdAt: string;
}

export interface SampleInput {
  title: string;
  description?: string;
  category: string;
  gameType?: string;
  budget?: number;
  previewImageUrl?: string;
  previewVideoUrl?: string;
  tags?: string;
}

export interface SampleUpdate {
  title?: string;
  description?: string;
  category?: string;
  gameType?: string;
  budget?: number;
  previewImageUrl?: string;
  previewVideoUrl?: string;
  tags?: string;
}

export interface LikeStatus {
  liked: boolean;
  likeCount: number;
}

export interface FavoriteStatus {
  favorited: boolean;
}

export interface FollowStatus {
  following: boolean;
  followerCount: number;
}

export interface Comment {
  id: number;
  sampleId: number;
  userId: number;
  content: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
}

export interface Request {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  gameType?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  referenceImageUrl?: string | null;
  status: string;
  /** @nullable */
  username?: string | null;
  createdAt: string;
}

export interface RequestInput {
  title: string;
  description?: string;
  category: string;
  gameType?: string;
  budget?: number;
  deadline?: string;
  referenceImageUrl?: string;
}

export interface DashboardStats {
  totalSamples: number;
  totalLikes: number;
  totalComments: number;
  totalViews: number;
  pendingSamples: number;
  approvedSamples: number;
}

export interface HomeStats {
  totalCreators: number;
  totalSamples: number;
  totalCategories: number;
}

export interface AdminStats {
  totalUsers: number;
  totalCreators: number;
  totalSamples: number;
  pendingApprovals: number;
  totalLikes: number;
}

export interface RejectInput {
  reason?: string;
}

export type Logout200 = {
  message: string;
};

export type GetCreatorsParams = {
search?: string;
category?: string;
limit?: number;
offset?: number;
};

export type GetTopCreatorsParams = {
limit?: number;
};

export type GetSamplesParams = {
category?: string;
game?: string;
budget?: number;
search?: string;
creatorId?: number;
status?: string;
limit?: number;
offset?: number;
};

export type GetTrendingSamplesParams = {
limit?: number;
};

export type GetRecentSamplesParams = {
limit?: number;
};

export type GetRequestsParams = {
category?: string;
status?: string;
limit?: number;
offset?: number;
};

export type AdminGetSamplesParams = {
status?: string;
limit?: number;
offset?: number;
};

export type AdminGetUsersParams = {
search?: string;
role?: string;
limit?: number;
offset?: number;
};