# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Only Creators - Gaming Creator Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # USERS
...
  /admin/samples:
    get:
      operationId: adminGetSamples
      tags: [admin]
      summary: Admin - list samples with filter
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sample"
  /admin/samples/{id}/approve:
    post:
      operationId: approveSample
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sample"
  /admin/samples/{id}/reject:
    post:
      operationId: rejectSample
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sample"
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Admin - list users
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/ban:
    post:
      operationId: banUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/creators/{id}/feature:
    post:
      operationId: featureCreator
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, email, password, accountType]
      properties:
        username: { type: string, minLength: 3 }
        email: { type: string }
        password: { type: string, minLength: 6 }
        accountType: { type: string, enum: [customer, creator] }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token: { type: string }

    # USER
    User:
      type: object
      required: [id, username, email, accountType, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        accountType: { type: string }
        avatarUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        role: { type: string }
        isBanned: { type: boolean }
        createdAt: { type: string }
...
    CreatorProfile:
      type: object
      required: [id, userId, displayName, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        displayName: { type: string }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        gmailAddress: { type: ["string", "null"] }
        discordUsername: { type: ["string", "null"] }
        servicesOffered: { type: ["string", "null"] }
        experienceLevel: { type: ["string", "null"] }
        portfolioUrl: { type: ["string", "null"] }
        socialLinks: { type: ["string", "null"] }
        verificationStatus: { type: string }
        level: { type: string }
        totalLikes: { type: integer }
        totalSamples: { type: integer }
        followerCount: { type: integer }
        isFeatured: { type: boolean }
        username: { type: ["string", "null"] }
        joinedAt: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Sample:
      type: object
      required: [id, creatorId, title, category, status, createdAt]
      properties:
        id: { type: integer }
        creatorId: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string }
        gameType: { type: ["string", "null"] }
        budget: { type: ["integer", "null"] }
        previewImageUrl: { type: ["string", "null"] }
        previewVideoUrl: { type: ["string", "null"] }
        tags: { type: ["string", "null"] }
        status: { type: string }
        likeCount: { type: integer }
        commentCount: { type: integer }
        isLiked: { type: boolean }
        isFavorited: { type: boolean }
        creatorName: { type: ["string", "null"] }
        creatorAvatarUrl: { type: ["string", "null"] }
        creatorVerified: { type: boolean }
        createdAt: { type: string }
...
    RejectInput:
      type: object
      properties:
        reason: { type: string }